RSDEV-260 Refresh Subsample preview after splitting
The Subsample's quantity field contains a piece of text that displays the
number of siblings. When a subsample is split, the number of siblings increases
and thus the label should be updated. This is done by triggering a re-fetch of
the subsample from the API, which in turn refreshes the UI.

This is a relatively minor bug caught whilst implementing RSDEV-260.