PRT-787 Remove split numerical field from Template create dialog
Because TemplateModel is a sub-class of SampleModel, the field for inputting
the number of splits to make of a subsample (or a sample's singular subsample
where there is only one) was visible in the template create dialog too.

This change fixes this issue by also checking that the record being operated on
is not a TemplateModel when deciding whether to show the split field.
Furthermore, CreateDialog.js has been refactored to use helper functions that
ensure the logic for determining the type of options to show is not duplicated,
limiting the probably of such a bug occurring in the future. Ultimately, this
file is too complicated and separate dialogs should be implemented for each
record type, rather than one with lots of conditional logic that increases the
chance of bugs.

Finally, a jest test is also added to check this specific case.