Fix tests that have become complected by shared global state
Commit 69ba63b7b5 introduced a cache for fetching deployment properties, so
that each only need be fetched once per page load. However, this fetch started
caching across tests, causing the first to be ran to pollute the running of the
rest that fetched a particular deployment property.

This change fixes this by allowing each test to run inside of its own cache
context, instead of using the default global one. It is unfortunate that anyone
who writes tests for components that use deployment properties will need to
remember to add the context; if we find this to be quite a sharp corner then we
may be able to change the implementation of useDeploymentProperty so that it
can detect when it is running inside of a node runtime and will not cache
anything.