RSDEV-243 Add download button to new gallery actions menu
On the current Gallery, this lives in the info panel but it is more akin to an
action that the user may perform on the file than it is a piece of metadata. As
such, it makes more sense to move it to a more intuitive place.

One thing to note, this uses the same icon as export and it may not be
immediately obvious to some users what the difference between download and
export is. Might be worth thinking about whether there is anything we can do to
make this distinction more obvious.