Fix accessibility warning in new Gallery's sidebar
We were setting the role of the List to "navigation" so that this semantic
information would be exposed to accessibility technologies like screen readers,
but by applying this role to the list wrapper we violated the rule that all
<li> tags must have a parent that is either an <ol> or a <ul>.

This change removes the custom role from the List to fix this violation and
introduces a new wrapper div with the "navigation" role to satisfy the
accessibility technologies. This has the added advantage of meaning that there
is only one navigation element that wraps both lists, which better describes
the fact that both lists perform the same type of navigation.