RSDEV-279 Additional subsample count label (#60)
This change adds an additional label on the subsample form informing the user of how many siblings the subsample has.

Currently, the subsamples quantity field provides a link to the search for all of the parent sample's subsamples, with a link text that specifies the number of subsamples minus one -- the number of siblings the subsample has. Given that the quantity field is in the details section, it may appear below the fold when the vertical dimension of the viewport is low or it may not be visible due to the details section being collapsed.

This change duplicates this label by adding it to the overview section's parent sample field also, where it is more likely to be seen at first glance.