PRT-683 Cancel batch edit completely
If whilst batch editing `uiStore.dirty` was true, i.e. the user had editing
some of the fields then `cancel` would not be called and the batch form field
would reset its state but it would not close to return to wherever the user was
before starting the batch editing. This change fixes this by ensuring that
`cancel` is always called, once any changes have been discarded.

This does not affect normal editing as calling `confirmDiscardAnyChanges` would
set the record back into `preview` mode and would re-fetch the current state of
the record from the backend. In the process of doing so, the editing operation
would be completely cancelled.