PRT-696 Update container contents after deleting
Not sure how this has not been a more pressing bug, but if you delete the
contents of a location in grid or image view the location was not updating, the
contents just stayed there. This change refreshes the container to ensure that
the UI keeps to date with the state.