Drag-and-drop cancel button
Before, the content being dragged was cut off by the edges of the scrolling box
so that if you dragged the card or tree node up to the breadcrumbs the preview
would vanish. This change uses `position: "fixed"` and z-index to ensure that
the content being dragged is always rendered above everything else.

However, by making the card/tree node position-fixed, it is removed from the
flow of the page and the adjacent content moves up to take its place. This
makes it harder to cancel the move operation as it is no longer possible to
simply move the item back to where it came from. As such, this change adds an
explicit cancel button: a floating-action button that appears whenever the
drag-and-drop operation is in progress.