RSDEV-282 By default, hide all but the most recent two notes on subsamples (#64)
If subsamples have lots of notes then it could make the UI quite
cumbersome, especially where the subsample notes are now shown on the
sample page they are no longer at the very foot of the page as they are
on the subsample page.

This change hides all but the two most recent notes by default, with the
rest hidden behind a "Show more" button. The new note field is placed at
the top of the list, where it appears before the most recent note and is
readily available even when the list of notes is long.

This change also improves the aesthetics of the notes somewhat,
utilising the theme colour of the subsamples to make it clear that the
notes are associated with the physical quantity of the sample, and not
the grouping that forms the whole sample.