Update VM_global_library.vm (Typo in email template)
Fixed a small error I noticed on password reset emails when it provides the IP address that requested the password reset, along with a link to see more info about that IP address.

There was a typo (used to be ip-adress.com), right now this still redirects to ipaddress.com, but in the future it might not. We don't know who might be Typosquatting this incorrect url in the future.

Furthermore, ipaddress.com has a lot of ads, so again, not a nice experience for a RSpace user who has clicked on it.

I've replaced ip-adress.com with ip-lookup.net.