PRT-790 Search navigations make left panel visible
Whenever a navigation is made to /inventory/search and is applied to the
page-wide search then the left panel should make visible so that on mobile a
user is sent back to the search results. Rather than doing it all over the
place, now that we have MainSearchNavigationContext we can do it in just one
place.