RSDEV-311 Make Inventory permalink parsing more robust
If the user accidentally enters an invalid value as an Id in a permalink they
are currently getting an incomprehensible Java error. With this change, we
attempt to correct their mistake if they enter a Global Id in place of an Id,
or else we redirect to the root of the application if a more helpful error
message.