RSDEV-304 Update subsample details when search changes (#77)
When the subsample search within the sample form changes (e.g. the user
searches for a particular subsample using the query string field), the
subsample details panel should update so that we're not showing the
details of subsample that are no longer in the table.

This is somewhat complicated by the next and previous buttons below the
details panel as they manipulated the search when the currently shown
subsample is up against a page boundary. The finer details of this
complication are in the code comments.