RSDEV-120 Migrate Argos dialog to accented theme (#78)
This means the Argos dialog for importing DMPs into the Gallery has
consistent styling with the new Gallery, utilising an accented theme to
mimic the Argos branding. Moreover, it supports the high contrast mode
that comes with the theme. The dialog also now uses MUI's DataGrid,
reducing the numbers of places that our more limited adjustable-column
table is used. This change applies to both the existing and new gallery.