PRT-791 Add error handling to fetching tags in export dialog
This change improves the error handling of calls to `/exportRecordTagsArchive`
in two ways.

First, a `.catch` clause is add to the promise so that should the response code
be anything but a 200 then the Add Tag button will not become inoperable, but
instead simply show no auto-suggested tags. By itself, that would actually
solve this bug as although the server is responding with a 200 with a payload
that describes an error when attempting to get the tags associated with a
Gallery file, attempting to use that error object as an array of string will
fail when we do `data.flatMap`.

However, this approach to the catching of errors that come in a 200 response is
not especially obvious. Instead, this change also incorporates richer parsing
logic for decoding the payload associated with the response. Instead of
assuming that the response will always be an array of strings, the response
object is parsed and should it not fit this shape (as it will not when fetching
the tags of a Gallery file) then again an empty array of tags is returned and
the "Add Tag" button remains operable.