RSDEV-120  Migrate DMPTool dialog to accented theme (#79)
This means the DMPTool dialog for importing DMPs into the Gallery has
consistent styling with the new Gallery, utilising an accented theme to
mimic the DMPTool branding. Moreover, it supports the high contrast mode
that comes with the theme. The dialog also now uses MUI's DataGrid,
allowing the user to choose additional columns of data. This change
applies to both the existing and new gallery.

This change also includes keyboard controls for selecting a row of the
table in the Argos dialog as that was not included in its recent change.