PRT-792 Improved error handling when adding items to baskets
Its not clear what the `await this.getItems();` call was intended for, but it
was resulting in a second 404 error when attempting to add an item to a deleted
basket. The other change is simply to ensure that the actual network call to
`/addItem` reports a more detailed error message than simply 'there was a 404'.