Move adjustable column files to Inventory/components
Since commit b46f2bcb43 replaced the usage of the adjustable column with
DataGrid, the only places the adjustable column is used is in Inventory. As
such, we can move the component there, given DataGrid is to be used for all
future use cases of such functionality. Hopefully at some point we can replace
the usage in Inventory too.