PRT-793 Number field should use decimal inputMode
This is because browsers are not required to give the user the means to input a
dash, and thus negative numbers, when the inputMode is "decimal" or "numeric";
and this includes iOS Safari. This is unfortunate as those inputModes present
the user with a numerical keypad, like a calculator or phone app, whereas
"text" is the standard qwerty keyboard.