Pure ArrayUtils functions should take $ReadOnlyArrays
Many of the functions in ArrayUtils are pure versions of the methods on
Array.prototype, returning new arrays rather than mutating the passed one.

By changing the type from Array to $ReadOnlyArray, this change ensures that the
code does not mutate the passed arrays and communicates to the calling code
that these functions are indeed pure with respect to the passed array.

As such, Flow is able to perform greater reasoning about the behaviour of the
code and thus we can avoid the need for a suppression in DefaultValueField