PRT-794 Define GeoLocationPolygon class to enforce bounding region (#93)
The first and last points of a GeoLocation polygon must be the same to
enclose a region. This invariant was not sufficiently being enforced and
so this change defines a new class and accompanying interface to ensure
all code that creates and modifies GeoLocation polygons enforces this
invariant.

Further changes were necessary because the GeoLocation classes could not
be used on the public page due to transitive dependencies on AuthStore,
which this change fixes with a little dependency injection. This makes
`publicData` unnecessary.