Fix import of types in IdentifierModel
This is a fix for commit e88c1de311 which amended the IdentifierModel class for
use on the public page, where the user is not authenticated, by passing all of
the global state via dependency injection. As such, we do not want to import
the class InvApiService, as that would trigger the import issues, but rather
simply its type to be able to type the parameters that are being passed in via
that injection.