Fix selected section indicator offset on mobile on new gallery page
The sidebars on the new gallery page have an indicator for the selected section
that animates up and down. The offset for this indicator is calculated based on
the height of the menu items, so that the animation always works even when the
page is zoomed.

On mobile, this offset was not being calculated correctly as sectionRefs, an
array of the refs for each of the menu items, was not initialised at the point
that the initial offset was being calculated as the sidebar was not in the DOM.

This change ensures that the sidebar, and thus the menu items with the refs, is
always in the DOM and that the offset is recalculated when these refs change
(e.g. due to a re-render) and thus the height of each menu item may change.