RSDEV-325 Use UI preferences to persist UI state in React-land (#102)
There are various small bits of the UI that would be good to persist
across user sessions: sort ordering, views, draft text.

This change adds a custom react hook for using bits of persisted state
in the same way that React.useState is used. The state is stored with
its timestamp so that in future we could implement a retention
policy/flush the oldest state when necessary. Finally, this change uses
this new hook to provide persistence for:
  * New Gallery's sorting
  * New Gallery's view
  * Inventory's form sections