Remove width limit of tree view on new gallery page
This maxWidth means that on mobile-sized viewports the tree only takes up half
of the horizontal space available to it, resulting in tree nodes with long file
names needlessly wrapping. As such, this change removes it.

The maxWidth was added by commit c1023b92f6, specifically commit 3679e155a8, as
part of the change to add a cancel button to the drag-and-drop functionality. I
don't recall the exact reason for why I added it, and removing it doesn't
appear to prevent the functionality from working correctly. Perhaps my initial
intention was to provide a region of the file listing where users could drop
the operation without it triggering a move, but in the end felt that an
explicit cancel dropzone would be more discoverable; I simply don't recall.

In any case, this is a key fix for smaller viewports, so if this change does
indeed end up causing issues then a different approach will be required.