PRT-804 Make displaying IGSN Dates more robust
Prior to commit e88c1de311, `d.value` was an ISO 8601 string directly from the
backend, the `toString()` call did nothing but satisfy flow, and the resulting
string rendered was of the form YYYY-MM-DD. With the application of
IdentifierModel to the public page, `d.value` became a Date (as the type would
imply), `toString()` converted it to a string specific to the user's locale and
thus it didn't have a 'T' in the right place. In the case where it didn't have
'T' at all, nothing would be rendered.

This change simply makes the rendering code more robust by not relying on
string munging. The rest of the change is an additional unit test to verify
that the fix works and some changes to the types to accurately reflect the
actual data as returned by the server.