RSDEV-308 Always offer to delete now empty samples (#110)
If all of the subsamples of a sample, where the sample had more than
one, were being simultaneously deleted then the info toast that offers
to delete the sample too was not being shown. This is due to cascading
nature of the state returned by the bulk endpoint, and the cache of
object initialisation of the MemoisedFactory. By reversing the list and
ensuring that the factory sees the final state first, coupled with a
filter to make sure we don't show the same toast multiple times, this is
easily fixed.