Remove excessive Promise.all
There doesn't appear to be any reason why these two sets of network calls have
to happen in series. `newFiles` was an array of response objects, which
`Promise.all` will treat as if it were wrapped in `Promise.resolve`.

The additional Promise.all was added by commit 800513ffdc, but its not clear
from the code, nor the PR, nor the JIRA ticket what the justification was for
the additional `Promise.all`