RSDEV-286 Add an info panel to the new gallery page (#113)
This replicates much of the same behaviour as the info panel on the
existing gallery, showing metadata and providing a primary action to
preview or edit the selected document. The main interesting addition is
a floating panel that can be swiped up on mobile, as the new gallery is
mobile friendly.