Rsdev 292 inventory thumbnail performance (#115)
### Background
Loading an inventory record can be slowed down by the front-end
requesting multiple of the same images. On loading an inventory record,
various images can be loaded:
	- main container image
	- sample image(s)
	- sub-sample image(s)
	- thumbnail versions of each of the above

In many cases the image is the same (images can inherit from their
parent e.g. all sub-samples use parent sample image unless a different
one is specified). We load (and store) these separately though, keyed on
`InventoryRecord` (Container, Sample or Sub-sample) ID. The images are
retrieved via multiple endpoints e.g.
	- `/samples/{sampleId}/image/{timestamp}`
	- `/samples/{sampleId}/thumbnail/{timestamp}`
	- `/subsamples/{sampleId}/image/{timestamp}`
	- `/subsamples/{sampleId}/thumbnail/{timestamp}`
- (`{timestamp}` in each of the above examples gets ignored by the
back-end)

These multiple endpoints reduce the chance to cache what is often the
same image for each of those items.

### Changes
`FileProperty` now contains a column `contentsHash`, which gets
populated when a new image is added. Any `FileProperty`, keyed on
`username` and `contentsHash`, will be re-used if the same image is
being saved.

A new endpoint `/inventory/file/image/{contentHash}` has been added and
is used where we generate HATEOAS links for an `InventoryRecord`, which
tells the front-end the url to retrieve the images(main image, thumbnail
image, and in the case of container location image) from, meaning any
request from the front-end for the same image by the same user uses the
same endpoint for improved cache-ability.

A liquibase migration has also been added to add the new
`FileProperty.contentsHash` column, and populate that column with the
hash of each the existing inventory files.

---------

Co-authored-by: matthias <matthias@researchspace.com>