PRT-805: Make sure we hash the apiKey with `SHA256` (instead of `BCrypt`)
The `BCrypt` is not ideal to be used to crypt apiKeys since it is not indexable on the database column. So in order to retrieve the user we were scanning all the list of users (O(n) complexity) instead of querying an indexed column of the database that would have had a constant complexity.