Send notification to user when async export fails (#130)
When an export e.g. export all users work fails, the user wasn't being
notified of the failure.

I wrapped the export generation attempt in a try-catch, and when there
is an exception it gets logged, and an export failure notification is
generated. Other types of export get handled in this way too.

The failure case can be recreated by setting
`src/main/resources/deployments/defaultDeployment.properties` property
`archive.folder.location` to an non-writable location.