PRT-816 Improve performance of useSearchParamState
For some reason, this code was causing a significant performance issue on the
new gallery page where the whole page was re-rendering multiple times per
second causing significant CPU usage. The useEffect was triggering over and
over due the searchParams object coming react router, which would update the
useState, which would re-render. I don't know whether the act of re-rendering
itself was causing the cycle (i.e. an infinite loop), whether the library was
firing off a new object multiple times a second, or something else entirely.

I think this change results in the same behaviour, only without the need for
the useEffect and the useState as useSearchParams already exposes a setter with
the same interface as useState. Calls to setSearchParams should result in this
component re-rendering, the state being re-calculated with the fallback value,
and the rest of the component tree re-rendering with the new state. Similarly,
if the fallback value is changed by the parent component then this component
should re-rendering, the state recalculated, and the rest of the component tree
re-rendering. However, whatever was causing the repeated re-renderings per
second should now not be happening.