PRT-821 Allow re-opening Aspose window
This change removes the `file` state from CallableAsposePreview and the need
for the useEffect, by instead just passing the state through to
CallablePdfPreview. By doing so, when the Pdf Preview's dialog is closed its
state is cleared, allowing for the state to be re-set for the same file and the
dialog re-opened.