Use user-event in jest tests (#159)
This change actually starts using @testing-library/user-event in the
jest tests, despite it always having been in the package.json. This
change includes type definition for the accompanying library as well as
type definitions for the part of @testing-library/react that we use.
These override the types downloaded from flow-types, allowing use to
type custom queries correctly. This change also migrates a couple of
tests from fireEvent to userEvent, [the reasons for doing so explained
here](https://testing-library.com/docs/user-event/intro#differences-from-fireevent).
Finally, this change removes usages of calling `click` on the found
element, which doesn't trigger a react re-rendering without being
wrapped in a call to `act` -- a foot-gun waiting to go off in many
places.