RSDEV-314 Convert Inventory create dialog into a wizard (#145)
This change converts the series of radio buttons with ad hoc fields into
a structured wizard, simplifying the code and streamlining the user
experience ahead of the addition of a new mechanism for creating
subsamples.

Most notably, this change removes the global create store, instead
attaching generic bits of state to the Search model and passing the
current search around when determining which locations of containers can
be selected. This has the benefit of reducing the amount of global state
and making it so that the components can be reused in various parts of
the UI with the behaviour dependent on which search is in scope of that
part of the component tree.

The create dialog has been entirely re-implemented to determine which
options and steps to display in the wizard based on the implementation
of InventoryRecord that it is passed, rather than relying on lots of
conditional logic in the dialog component. This way, we can continue to
add new functionality for creating new records in various different ways
in a way that scales more linearly.