RSDEV-314 Create subsamples from existing samples (#167)
Adds a new option to the sample create dialog, allowing for new
subsamples to be created without reference to an existing subsamples
i.e. without splitting or duplicating subsamples.

Much of this change deals with the repercussions to all of the unit
tests of using the unit store to display a label in the create dialog of
how much additional quantity the user wishes to create of the sample in
question.