Correct calls to moveGalleriesElements endpoint
Calls to `moveFiles().to()` were always moving the selected files to the root
of the gallery section as the function was not checking the union type and
dispatching on `moveFiles().toDestinationWithFolder()` when the destination is
a folder. This resulted in a drag-and-drop operations always moving the dragged
file to the root of the gallery section even when the file had been dropped
over another folder. This bug didn't affect the move dialog as it calls
`toDestinationWithFolder` directly

The check for snippet folders should then live in `toDestinationWithFolder()`
where all attempts to move anything into a snippets folder will be caught.