RSDEV-391 Remove drag-and-drop panel from new gallery; supporting dropping anywhere instead (#172)
The panel that was appearing to act as a dropzone for
dragging-and-dropping files from outside of the browser and into the new
Gallery was not obvious for some absent-mindedly dragging files in. Any
drag-and-drop event that concluded outside of the panel was silently
ignored and user may not notice that they did in fact not import the
files.

Increasing the size of the panel to cover the whole page was not an
option as we would lose the ability to drag-and-drop files into a
specific folder. Therefore, the background of the card view was instead
turned into the dropzone, with a blue outline to indicate this.