RSDEV-389 Fix MoveDialog jest test
Commit 7e2863faad added CallablePdfPreview to TreeView so that double tapping a
PDF in TreeView would open a preview of it. Given TreeView is used by the
MoveDialog -- even though PDFs are never listed in its tree -- this meant that
MoveDialog had an indirect dependency on CallablePdfPreview.

Now, CallablePdfPreview uses some syntax that only works at runtime in the
browser so to make the MoveDialog unit tests continue to work we mock
CallablePdfPreview.