Fix bug with open menu item in new gallery actions menu
If the user selected a folder the open menu item would get added to the actions
menu but if they then changed the selection to something that wasn't a folder,
or expanded the selection to contain more than a single folder, then the open
menu item would remain when it should be removed.

This appears to have been due to the `key` attribute. I've been working under
the assumption that when eslint complains about a Node inside of a `map` call
from an Optional or Result not having a key it is simply looking for the name
`map` and not doing any sophisticated static analysis, and that, as such, I can
pass `null` and it will simply be ignored. It seems, however, that this key is
not ignored by the react runtime and does have some unexpected side effects. I
don't fully understand the full details of why a key of `null` results in the
menu item not being removed from the menu, but it seems that using `null` as a
key is not something I should do in the future.