RSDEV-390 Fix cursor inside photoswipe
When the ImagePreview, as provided by photoswipe, is open the cursor adjusts.
If the image can be zoomed in then "zoom-in" is used, if the image has been
zoomed in then "grab" and "grabbing" are used as the user scrolls around, and
if the image cannot zoom in that "zoom-out" is used for some very strange
reason even though tapping causes the gallery to close.

This change reimplements the cursor styles to ensure that if an image cannot be
zoomed in then the default cursor style is applied.