Holding a gallery folder should not attempt to move it into itself
If the user accidentally taps and holds a file card of a folder, the code
should not attempt to move the folder into itself (which the backend will
silently refuse to do). Instead we should check that the filesBeingMoved does
not include the dropped destination, so that we don't end up showing the user a
success toast that says their folder was successfully moved, when it wasn't
moved anywhere.