RSDEV-339 Replace image editor in Inventory (#189)
This change replaces the image editor implemented with
`@toast-ui/react-image-editor` -- which is no longer supported -- with
one implemented mostly by us with use of the much smaller
`react-image-crop` library. `react-image-editor` is not removed from the
codebase yet, as it is also used by the Gallery, but once the Gallery
has been migrated to the new UI code we will be able to remove this
deprecated dependency, which will mean we will no longer have to pass
`--force` to `npm install`.