PRT-826 Move DialogTitle out of DialogContent
Commit 2d299e650c introduced the bug by removing the margin from all
DialogTitles. This margin is however need for all of the dialogs which don't
have an app bar header and the reason that the use of DialogTitles in the
dialogs with the app bar header resulted in excess margin is because the
DialogTitles were being rendered inside of DialogContent. As such, this change
moves the DialogTitles and reinstates the margin.