PRT-831 iRODS dialog: fix scrolling of dialog content
The `<form>` tag sitting between the `<Box>` and the `<DialogTitle>`,
`<DialogContent>`, and `<DialogActions>` broke the flexbox-based layout that
causes the content to become a scroll container when there is insufficient
vertical space. By having the `<Box>` **be** a `<form>` we get the best of both
worlds, although I can't say I can recall why we're using a '<form>` here to
begin with given that we're not using its onSubmit event to trigger a
submission when the enter key is pressed. Possibly it is just for the username
and password to be populated by the browser. In any case, this change fixes the
scrolling issue.