Enforce comments on exports (#203)
To encourage better documentation practices, this change enforces all JS
exports from one module to the rest of the codebase to have a comment.
Ideally this comment should explain anything that the calling code from
a different module needs to know about the implementation details of the
export e.g. that it can throw an error, the behaviour of edge-cases,
performance caveats, etc.