Improve responsiveness when tapping FileCard
Drag-and-drop was causing a lot of re-renders when the user taps a FileCard,
resulting in the new selection not showing up in the UI for a few seconds. This
makes the UI feel very slow and janky. This was made extra worse by the fact
that the TouchRipple animation doesn't play until after the two drag-and-drop
renders.

This change reduces the amount of re-rendering by preventing drag-and-drop from
triggering until 500ms have elapsed. This still results in a second or so of
render time, but with the touch ripple animation playing at the same time this
is re-render, the UI no longer feels nearly as slow.