Remove accent checkbox and radio colours
They automatically pull the theme colours and by setting the `fill` we end up
overriding the disabled state that applies a grey colours instead, therefore
it's best to just not set the `fill` colours this way. Difficult to say why
this was added originally, it's possible that we weren't setting the palette
correctly before.