Improve new Gallery's listing refresh logic (#208)
When the user performs an action such as creating folders, uploading
files, or moving, deleting, or duplicating existing files, the listing
is refreshed to show them the new state of the current folder/gallery
section that they're viewing. This refresh logic was only refreshing the
first page of results so that if they had loaded several pages and then
performed an action, they would have manually reload all those pages
again. This was particularly annoying if one wanted to perform multiple
operations on a single selection.

This change refetches all of the pages that have been loaded, near
seamlessly updating the UI, with the scroll position remaining in place.
One small caveat is when some of the selected files are no longer in the
listing because the addition of some new files has caused the some of
the selection to be shifted into the next page, which has not been
loaded. In that circumstance we do still clear the selection.