RSDEV-434 Alter layout of new Gallery page (#219)
To make the important information more obvious and show how information
is related through visual hierarchies, this change tweaks how the new
gallery is organised somewhat. It places the listing controls above both
the listing and the info panel, changes the breadcrumbs into a series of
chips/pills, and alters the typographical styles of the info panel
subheadings.